Consolidating simulation constructor

+++ src/edu/gcc/results/Results.java
@@ -3,6 +3,13 @@ package edu.gcc.results;
 import java.util.List;
 import java.util.ArrayList;
 
+/**
+ * This class is a storage container for the results of a simulation. This
+ * includes statistics like average time per order, average distance per order,
+ * etc.
+ * 
+ * @author Ethan Harvey, Zack Orlaski, Luke Donmoyer
+ */
 public class Results {
 	// Chart 1
 	private List<Long> timePerOrder;
@@ -10,48 +17,51 @@ public class Results {
 	private long longestTime;
 	// Chart 2
 	private List<Integer> ordersPerTrip;
-	private int averageOrdersPerTrip;
+	private double averageOrdersPerTrip;
 	// Chart 3
 	private List<Long> distancePerTrip;
 	private long averageDistancePerTrip;
-	
+
 	private String simType;
-	
+
 	public Results(
 			List<Long> timesPerOrder,
 			List<Integer> ordersPerTrip2,
 			List<Long> distancePerTrip2,
 			String simType
-			) {
+	) {
 		this.timePerOrder = timesPerOrder;
 		this.ordersPerTrip = ordersPerTrip2;
 		this.distancePerTrip = distancePerTrip2;
 		this.simType = simType;
-		
+
 		// Set average time per order
 		this.averageTimePerOrder = 0;
 		this.longestTime = 0;
-		for(Long time : this.timePerOrder) {
+		for (Long time : this.timePerOrder) {
 			this.averageTimePerOrder += time;
-			if(time > this.longestTime) {
+			if (time > this.longestTime) {
 				this.longestTime = time;
 			}
 		}
-		this.averageTimePerOrder = this.averageTimePerOrder / this.timePerOrder.size();
-		
+		this.averageTimePerOrder = this.averageTimePerOrder / this.timePerOrder
+				.size();
+
 		// Set average orders per trip
 		this.averageOrdersPerTrip = 0;
-		for(int orders : this.ordersPerTrip) {
+		for (int orders : this.ordersPerTrip) {
 			this.averageOrdersPerTrip += orders;
 		}
+    
 		this.averageOrdersPerTrip = this.averageOrdersPerTrip / this.ordersPerTrip.size();
-		
+
 		// Set average distance traveled per trip
 		this.averageDistancePerTrip = 0;
-		for(Long distance : this.distancePerTrip) {
+		for (Long distance : this.distancePerTrip) {
 			this.averageDistancePerTrip += distance;
 		}
-		this.averageDistancePerTrip = this.averageDistancePerTrip / this.distancePerTrip.size();
+		this.averageDistancePerTrip = this.averageDistancePerTrip
+				/ this.distancePerTrip.size();
 	}
 
 	public List<Long> getTimePerOrder() {
@@ -78,11 +88,11 @@ public class Results {
 		this.ordersPerTrip = ordersPerTrip;
 	}
 
-	public long getAverageOrdersPerTrip() {
+	public double getAverageOrdersPerTrip() {
 		return averageOrdersPerTrip;
 	}
 
-	public void setAverageOrdersPerTrip(int averageOrdersPerTrip) {
+	public void setAverageOrdersPerTrip(double averageOrdersPerTrip) {
 		this.averageOrdersPerTrip = averageOrdersPerTrip;
 	}
 
@@ -109,13 +119,12 @@ public class Results {
 	public void setAverageDistancePerTrip(long averageDistancePerTrip) {
 		this.averageDistancePerTrip = averageDistancePerTrip;
 	}
-	
+
 	public String getSimType() {
 		return simType;
 	}
-	
+
 	public void setSimType(String simType) {
 		this.simType = simType;
 	}
-	
 }
